Refactor: Extract LinkPopup and clean up dead code
- Extract  component to  for better modularity.
- Remove unused  prop and logic from  and associated components.
- Remove dead  function from .
